refactor to make more clear where code for each model lives
