+++ apps/launch-summarize-app.py
@@ -38,4 +38,4 @@
 #
 # ###########################################################################
 
-!streamlit run apps/rebrief.py  --server.port $CDSW_APP_PORT --server.address 127.0.0.1
+!streamlit run apps/summarize-app.py  --server.port $CDSW_APP_PORT --server.address 127.0.0.1
