+++ summa/st_model_wrappers.py
@@ -39,24 +39,10 @@
 # ###########################################################################
 
 import attr
-import transformers as trf
 
-from summa.models.classic_extractive import (
-    SentenceTextRank, 
-    build_classic_nlp_pipeline,
-    build_trf_nlp_pipeline, 
-    classic_summary, 
-    sentence_summary_upgrade
-)
-from summa.models.neural_extractive import (
-    SentenceBertClass, 
-    load_neural_extractive_model,
-    summarize
-)
-
-# this is the current default when loading the HF summarization pipeline
-# making it explicit here to reduce ambiguity.
-ABSUM_MODEL = "sshleifer/distilbart-cnn-12-6" 
+from summa.models import neural_extractive as ne
+from summa.models import neural_abstractive as na
+from summa.models import classic_extractive as ce
 
 @attr.s()
 class SummarizationModel(object):
@@ -104,41 +90,10 @@ class SummarizationModel(object):
     def __hash__(self):
         return self.name
 
-
-def load_abstractive_model(model_name = ABSUM_MODEL):
-    return trf.pipeline("summarization", model=model_name, tokenizer=model_name)
-
-def abstractive_summary(text, model):
-    try:
-        output = model(text, return_tensors=False, clean_up_tokenization_spaces=True)
-        summary = output[0]['summary_text']
-    except IndexError:
-        # the input text is too long. Need to break it up. 
-        paragraphs = text.split("\n")
-        paragraphs = [p for p in paragraphs if p]
-        summary = []
-        for paragraph in paragraphs:
-            try:
-                output = model(paragraph, return_tensors=False, clean_up_tokenization_spaces=True)
-                summary.append(output[0]['summary_text'])
-            except IndexError:
-                # if a paragraph is STILL too long, split further
-                sentences = paragraph.split(".") 
-                # TODO: need to generalize this because these chunks might be too long
-                chunks = 2 
-                segment_size = int(len(sentences)/chunks)
-                while sentences:
-                    segment = ". ".join(sentences[:segment_size])
-                    sentences = sentences[segment_size:]
-                    output = model(segment, return_tensors=False, clean_up_tokenization_spaces=True)
-                    summary.append(output[0]['summary_text'])
-        summary = "\n".join(summary)
-    return summary
-
 abstractive = SummarizationModel(
     name = "abstractive",
-    load = load_abstractive_model,
-    summarize = abstractive_summary,
+    load = na.load_abstractive_model,
+    summarize = na.abstractive_summary,
     display_name = "Neural Abstractive",
     description = "HuggingFace Summarization Pipeline.\n\n HuggingFace \
     provide models that perform _abstractive_ summarization. \
@@ -158,8 +113,8 @@ abstractive = SummarizationModel(
 
 modern_extractive = SummarizationModel(
     name = "modern_extractive",
-    load = load_neural_extractive_model,
-    summarize = summarize,
+    load = ne.load_neural_extractive_model,
+    summarize = ne.summarize,
     display_name = "Neural Extractive",
     description = "Fine-tuning SentenceBERT.\n\n For this model \
     we train a Transformer to perform _extractive_ rather than _abstractive_ summarization. \
@@ -179,8 +134,8 @@ modern_extractive = SummarizationModel(
 
 classic_extractive = SummarizationModel(
     name = "classic_extractive",
-    load = build_classic_nlp_pipeline,
-    summarize = classic_summary,
+    load = ce.build_classic_nlp_pipeline,
+    summarize = ce.classic_summary,
     display_name = "Classic Extractive",
     description = "TextRank.\n\n TextRank is a classic graph-based ranking \
     algorithm that computes the importance of a vertex given global information \
@@ -200,10 +155,10 @@ classic_extractive = SummarizationModel(
     and sentences containing the top phrases are extracted as a summary.",
 )
 
-upgraded_classic_extractive = SummarizationModel(
+hybrid_extractive = SummarizationModel(
     name = "hybrid_extractive",
-    load = build_trf_nlp_pipeline,
-    summarize = sentence_summary_upgrade,
+    load = ce.build_trf_nlp_pipeline,
+    summarize = ce.sentence_summary_trf,
     display_name = "Hybrid Extractive",
     description = "TextRank + SentenceBERT.\n\n This hybrid approach relies on the \
     same basic tenets of the \"Classic Extractive\" model but with a twist.\
